Рефакторинг IAccountService
