+++ HPM-System.IdentityServer/Services/AccountService/LoginResultData.cs
@@ -1,0 +1,10 @@
+﻿namespace HPM_System.IdentityServer.Services.AccountService
+{
+    public class LoginResultData
+    {
+        public string UserId { get; set; } = string.Empty;
+        public string Email { get; set; } = string.Empty;
+        public string PhoneNumber { get; set; } = string.Empty;
+        public string Token { get; set; } = string.Empty;
+    }
+}
